+++ CustomGameEngine/Shaders/g_bufferPBR.frag
@@ -3,7 +3,7 @@ layout (location = 0) out vec3 gPosition;
 layout (location = 1) out vec3 gNormal;
 layout (location = 2) out vec3 gAlbedo;
 layout (location = 4) out vec3 gARM; // Ambient, Roughness, Metalness
-layout (location = 5) out uint gPBRFLAG;
+layout (location = 5) out float gPBRFLAG;
 
 in VIEW_DATA {
     flat vec3 TangentViewPos;
@@ -167,5 +167,5 @@ void main() {
     gARM.g = Roughness;
     gARM.b = Metallic;
 
-    gPBRFLAG = uint(1);
+    gPBRFLAG = 1.0;
 }
